Apply transformer's tests
