+++ test/transformer/pandas/test_EvalTF.py
@@ -1,0 +1,9 @@
+from dtoolkit.transformer import EvalTF
+
+
+def test_work():
+    new_column = "double_value"
+    tf = EvalTF(f"`{new_column}` = `{feature_names[0]}` * 2")
+    result = tf.fit_transform(df_iris)
+
+    assert result[new_column].equals(df_iris[feature_names[0]] * 2)
