+++ test/transformer/pandas/test_ReplaceTF.py
@@ -1,0 +1,9 @@
+from dtoolkit.transformer import ReplaceTF
+
+
+def test_work():
+    tf = ReplaceTF({1: "a"})
+
+    result = tf.fit_transform(df_label)
+
+    assert result.isin(["a"]).any(axis=None)
